Simplify driver only test case
